// src/app/features/appointment-report/models/appointment-report.model.ts
export class AppointmentReportModel {
  Status?: string;
  MonthSet?: string;
  Date?: string;
  Doctor?: string;
  Rep?: string;
  City?: string;
  State?: string;
  Phone?: number;
  Lab?: string;
  LunchDate?: string;
  MonthScheduled?: string;
  POC?: string;
  Setter?: string;
  ConfirmationCallDate?: string;
  Notes?: string;
}